refactor(ui): split components surfaces
