+++ apps/desktop/src/main/__tests__/text-file-import.test.ts
@@ -324,7 +324,7 @@ describe('text file context import', () => {
     const preloadSource = await readFile(join(process.cwd(), 'src/preload/preload.ts'), 'utf8');
     const globalSource = await readFile(join(process.cwd(), 'src/global.d.ts'), 'utf8');
     const onboardingSource = await readFile(join(process.cwd(), 'src/renderer/OnboardingHero.tsx'), 'utf8');
-    const uiSource = await readFile(join(process.cwd(), '../../packages/ui/src/components.tsx'), 'utf8');
+    const uiSource = await readFile(join(process.cwd(), '../../packages/ui/src/composer.tsx'), 'utf8');
     const cssSource = await readFile(join(process.cwd(), 'src/renderer/maka-tokens.css'), 'utf8');
     const stylesSource = await readRendererContractCss();
     const textFilePromptBlock = mainSource.match(/async function importTextFilePrompt[\s\S]*?async function importTextFileIntoComposer/)?.[0] ?? '';
